Refactor predicates.
