+++ nmmo/task/__init__.py
@@ -2,4 +2,3 @@ from .game_state import *
 from .predicate import *
 from .task_api import *
 from .sampler import *
-from .base_predicate import *
