+++ nmmo/task/predicate/__init__.py
@@ -1,0 +1,1 @@
+from nmmo.task.predicate.core import Predicate
